i remove the migration for i made a rollback and changed the cellphone data type and either converted to bigint and i corrected the test

+++ db/migrate/20190828170142_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[5.2]
       t.string :first_name
       t.string :last_name
       t.string :email
-      t.integer :cellphone
+      t.bigint :cellphone
       t.text :address
 
       t.timestamps
